Complete memory allocator

+++ lab3/include/mini_uart.h
@@ -9,7 +9,7 @@ public:
     static void GetS(char* str);
     static void PutS(const char* str);
     static void PutS(const char* str, uint64_t length);
-    static void PutUInt32(uint32_t val);
+    static void PutUInt64(uint64_t val);
     static uint8_t GetCh();
     static void Send(uint8_t ch);
     static uint8_t Recv();
@@ -17,8 +17,8 @@ public:
 
 class IO {
 public:
-    IO operator<<(uint32_t val) {
-        MiniUART::PutUInt32(val);
+    IO operator<<(uint64_t val) {
+        MiniUART::PutUInt64(val);
         return *this;
     }
     IO operator<<(const char* str) {
